==> notebooks/01_data_split.ipynb <==
# Data Split

import pandas as pd
import numpy as np

# Load raw training dataset
raw_train_df = pd.read_csv('../data/raw/train.csv')
raw_train_df = raw_train_df.drop(columns=['id'])
raw_train_df

# Split into train/valid/test (80/10/10)

n_observations = len(raw_train_df)

valid_ratio = 0.1
test_ratio = 0.1
print(f'Train / Val / Test : {100*(1-valid_ratio-test_ratio)} / {100*valid_ratio} / {100*test_ratio}')

n_valid = int(n_observations * valid_ratio)
n_test = int(n_observations * valid_ratio)

rng = np.random.default_rng(123)   # For reproducible results

# Sample valid split
valid_split_df = raw_train_df.sample(n = n_valid, random_state=rng)
train_split_df = raw_train_df.drop(valid_split_df.index)

# Sample test split
test_split_df = train_split_df.sample(n = n_test, random_state=rng)
train_split_df = train_split_df.drop(test_split_df.index)

print('Size of training   split : ', len(train_split_df))
print('Size of validation split : ', len(valid_split_df))
print('Size of test       split : ', len(test_split_df))

# Export train/val/test split

train_split_df.to_csv('../data/processed/train_split.csv', index=False)
valid_split_df.to_csv('../data/processed/valid_split.csv', index=False)
test_split_df.to_csv('../data/processed/test_split.csv', index=False)
