==> notebooks/03_feature_engineering.ipynb <==
# Feature Engineering

import pandas as pd
import numpy as np
from sklearn.preprocessing import OrdinalEncoder, OneHotEncoder

import warnings
warnings.simplefilter("ignore", category=pd.errors.PerformanceWarning)   # For convinience

# Load cleaned train/val split and test dataset
train_df = pd.read_csv('../data/processed/cleaned_train.csv')   # Fit transformation only on the train split
valid_df = pd.read_csv('../data/processed/cleaned_valid.csv')
test_df = pd.read_csv('../data/processed/test_split.csv')

# Import EDA information
summary_cat_df = pd.read_csv('../data/processed/summary_cat.csv', index_col=0)
summary_cat_df

cat_vars = [var for var in train_df.columns if var.startswith('cat')]
cont_vars = [var for var in train_df.columns if var.startswith('cont')]

Things to notice:
- We should make sure the encoders and transforms should be fitted only by the train split
- Handle unseen data in val/test split

# Fequency encoding
counts = []
for cat in cat_vars:
    count = train_df[cat].value_counts()
    counts.append(count)

    train_df[f'{cat}_freq'] = train_df[cat].map(count)
    train_df[f'{cat}_log_freq'] = np.log1p(train_df[f'{cat}_freq'])     # take log to compress extreme counts
    train_df[f'{cat}_norm_freq'] = train_df[f'{cat}_freq'] / len(train_df)

    valid_df[f'{cat}_freq'] = valid_df[cat].map(count)
    valid_df[f'{cat}_log_freq'] = np.log1p(valid_df[f'{cat}_freq'])     # take log to compress extreme counts
    valid_df[f'{cat}_norm_freq'] = valid_df[f'{cat}_freq'] / len(train_df)

    test_df[f'{cat}_freq'] = test_df[cat].map(count).fillna(0)
    test_df[f'{cat}_log_freq'] = np.log1p(test_df[f'{cat}_freq'])     # take log to compress extreme counts
    test_df[f'{cat}_norm_freq'] = test_df[f'{cat}_freq'] / len(train_df)

train_df.head()

for cat, count in zip(cat_vars, counts):
    print(cat)
    print(count)
    print("")

# Ordinal Encoding (Much easier to handle unseen data)
ord_en = OrdinalEncoder(handle_unknown='use_encoded_value', unknown_value=-1)
ord_en.fit(train_df[cat_vars])
    
train_df[cat_vars] = ord_en.transform(train_df[cat_vars])
valid_df[cat_vars] = ord_en.transform(valid_df[cat_vars])
test_df[cat_vars] = ord_en.transform(test_df[cat_vars])

train_df.head()

# One Hot Encoding
# Only on low cardinality columns
low_card_cols = summary_cat_df[summary_cat_df['n_cats'] <= 3].index.to_list()

onehot_en = OneHotEncoder(handle_unknown='ignore', sparse_output=False)
onehot_en.fit(train_df[low_card_cols])

train_onehot_encoded = onehot_en.transform(train_df[low_card_cols])
valid_onehot_encoded = onehot_en.transform(valid_df[low_card_cols])
test_onehot_encoded = onehot_en.transform(test_df[low_card_cols])

train_onehot_df = pd.DataFrame(train_onehot_encoded, columns=onehot_en.get_feature_names_out(low_card_cols))
valid_onehot_df = pd.DataFrame(valid_onehot_encoded, columns=onehot_en.get_feature_names_out(low_card_cols))
test_onehot_df = pd.DataFrame(test_onehot_encoded, columns=onehot_en.get_feature_names_out(low_card_cols))

train_df = pd.concat([train_df, train_onehot_df], axis=1)
valid_df = pd.concat([valid_df, valid_onehot_df], axis=1)
test_df = pd.concat([test_df, test_onehot_df], axis=1)

train_df.head()

# Categorical groups numeric stats
for cat in cat_vars:
    group = train_df.groupby(cat)
    for cont in cont_vars:
        # Calculate global stats for unseen categories
        global_mean = train_df[cont].mean()
        global_med = train_df[cont].median()
        global_std = train_df[cont].std()
        
        train_df[f'{cat}_{cont}_mean'] = train_df[cat].map(group[cont].mean())
        valid_df[f'{cat}_{cont}_mean'] = valid_df[cat].map(group[cont].mean()).fillna(global_mean)
        test_df[f'{cat}_{cont}_mean'] = test_df[cat].map(group[cont].mean()).fillna(global_mean)

        train_df[f'{cat}_{cont}_med'] = train_df[cat].map(group[cont].median())
        valid_df[f'{cat}_{cont}_med'] = valid_df[cat].map(group[cont].median()).fillna(global_med)
        test_df[f'{cat}_{cont}_med'] = test_df[cat].map(group[cont].median()).fillna(global_med)

        train_df[f'{cat}_{cont}_std'] = train_df[cat].map(group[cont].std())
        valid_df[f'{cat}_{cont}_std'] = valid_df[cat].map(group[cont].std()).fillna(global_std)
        test_df[f'{cat}_{cont}_std'] = test_df[cat].map(group[cont].std()).fillna(global_std)

train_df.head()

# Numerics transformation

# Rank encoding for numerics
def fit_rank_transform(df, col):
    unique = sorted(df[col].unique())
    ranks = np.searchsorted(unique, unique) / len(unique)
    return unique, ranks

def transform_rank(values, unique, ranks):
    ''' Return the rank of value given the fitted rank transform (unique, ranks) '''
    return np.interp(values, unique, ranks)

# Winsorization (cap extremes)
def fit_winsor(df, col):
    l, u = df[col].quantile([0.01, 0.99])
    return l, u


for cont in cont_vars:
    # Log transform
    train_df[f'log_{cont}'] = np.log1p(train_df[cont])
    valid_df[f'log_{cont}'] = np.log1p(valid_df[cont])
    test_df[f'log_{cont}'] = np.log1p(test_df[cont])

    # Rank transform
    uniq, ranks = fit_rank_transform(train_df, cont)
    train_df[f'{cont}_rank'] = transform_rank(train_df[cont], uniq, ranks)
    valid_df[f'{cont}_rank'] = transform_rank(valid_df[cont], uniq, ranks)
    test_df[f'{cont}_rank'] = transform_rank(test_df[cont], uniq, ranks)

    # Winsorization
    l, u = fit_winsor(train_df, cont)
    train_df[f'{cont}_cap'] = train_df[cont].clip(l, u)
    valid_df[f'{cont}_cap'] = valid_df[cont].clip(l, u)
    test_df[f'{cont}_cap'] = test_df[cont].clip(l, u)

# Transform target loss
train_df[f'log_loss'] = np.log(train_df['loss'])
valid_df[f'log_loss'] = np.log(valid_df['loss'])
test_df[f'log_loss'] = np.log(test_df['loss'])
    
train_df

valid_df

## Export engineered dataframes

# Store in parquet format for efficiency
train_df.to_parquet('../data/processed/transformed_train.parquet', index=False)
valid_df.to_parquet('../data/processed/transformed_valid.parquet', index=False)
test_df.to_parquet('../data/processed/transformed_test.parquet', index=False)
