==> notebooks/04_LightGBM.ipynb <==
# LightGBM Modeling & Hyperparameter Tuning

import pandas as pd
import numpy as np
import lightgbm as lgb
import optuna
import mlflow, mlflow.lightgbm
from sklearn.metrics import mean_absolute_error

# Load preprocessed data
train_df = pd.read_parquet('../data/processed/transformed_train.parquet')
valid_df = pd.read_parquet('../data/processed/transformed_valid.parquet')

# Define target and features
target = 'log_loss'
X_train, Y_train = train_df.drop(columns=['loss', target]), train_df[target]
X_valid, Y_valid = valid_df.drop(columns=['loss', target]), valid_df[target]

print('Train shape : ', X_train.shape)
print('Valid shape : ', X_valid.shape)

# Prepare lightgbm dataset
#train_lgb = lgb.Dataset(X_train, label=Y_train)
#valid_lgb = lgb.Dataset(X_valid, label=Y_valid)

# Define Optuna task (with MLflow)

def objective(trial):
    
    params = {
        'objective': 'regression',
        'metric': 'mae',

        'num_leaves': trial.suggest_int('num_leaves', 64, 512, log=True),
        'max_depth': trial.suggest_int('max_depth', 6, 14),
        'min_data_in_leaf': trial.suggest_int('min_data_in_leaf', 200, 3000, log=True),

        'learning_rate': trial.suggest_float('learning_rate', 0.005, 0.05, log=True),

        'feature_fraction': trial.suggest_float('feature_fraction', 0.5, 0.9),
        'feature_fraction_bynode': trial.suggest_float('feature_fraction_bynode', 0.5, 1.0),
        'bagging_fraction': trial.suggest_float('bagging_fraction', 0.6, 0.9),
        'bagging_freq': trial.suggest_int('bagging_freq', 1, 10),

        'lambda_l1': trial.suggest_float('lambda_l1', 0.0, 10.0),
        'lambda_l2': trial.suggest_float('lambda_l2', 0.0, 50.0),

        'min_gain_to_split': trial.suggest_float('min_gain_to_split', 0.0,  5.0),

        'max_bin': trial.suggest_categorical('max_bin', [127, 255]),

        'boosting_type': 'gbdt',
        'verbosity': -1,
        'seed': 42,
    }

    train_lgb = lgb.Dataset(X_train, label=Y_train, params=params)
    valid_lgb = lgb.Dataset(X_valid, label=Y_valid, params=params, reference=train_lgb)

    with mlflow.start_run(nested=True):
        mlflow.log_params(params)

        model = lgb.train(
            params,
            train_lgb,
            valid_sets = [valid_lgb],
            num_boost_round=10000,
            callbacks = [
                lgb.early_stopping(stopping_rounds=200),
                lgb.log_evaluation(period=100),
            ],
        )

        y_pred = model.predict(X_valid, num_iteration=model.best_iteration)
        mae = mean_absolute_error(np.exp(Y_valid), np.exp(y_pred))  # Calculate MAE based on true loss
        mlflow.log_params(params)
        mlflow.log_metric('mae', mae)
        mlflow.log_metric('best_iteration', model.best_iteration)

    return mae           

# Run Optuna study

# Use mlruns folder for MLflow results
mlflow.set_tracking_uri("../mlruns")

# Set and start experiment
mlflow.set_experiment('lgbm_optuna_allstate')

study = optuna.create_study(
    direction = 'minimize',
    sampler = optuna.samplers.TPESampler(seed=42),
)

#study.optimize(objective, timeout=3*60*60)  # Run for 3 hours
study.optimize(objective, n_trials=20)

print('Best MAE: ', study.best_value)
print('Best params: ', study.best_params)

'''
Run 20 trials
Best MAE:  1152.6475239390618
Best params:  
{'num_leaves': 236, 
 'max_depth': 9, 
 'min_data_in_leaf': 203, 
 'learning_rate': 0.015748462517233382, 
 'feature_fraction': 0.7604063212686378, 
 'feature_fraction_bynode': 0.7966713343930665, 
 'bagging_fraction': 0.8106707103422641, 
 'bagging_freq': 7, 
 'lambda_l1': 2.721190423876234, 
 'lambda_l2': 18.412028624412162, 
 'min_gain_to_split': 0.17783197825737135, 
 'max_bin': 255}
'''


'''
# Run 40 trials
Early stopping, best iteration is:
[1400]	valid_0's l1: 0.416135
[I 2025-12-13 01:03:05,485] Trial 36 finished with value: 1140.1649585398006 and parameters: {'num_leaves': 404, 'max_depth': 8, 'min_data_in_leaf': 202, 'learning_rate': 0.02361306858113298, 'feature_fraction': 0.7211579800007945, 'feature_fraction_bynode': 0.9081132434036039, 'bagging_fraction': 0.8727967232517579, 'bagging_freq': 7, 'lambda_l1': 5.763583420124947, 'lambda_l2': 14.460054311753241, 'min_gain_to_split': 0.032912218537174215, 'max_bin': 127}. Best is trial 36 with value: 1140.1649585398006.

# Best params 
# MAE: 1140.1649585398006 
# trial 36 (40 trials)
{'num_leaves': 404, 
 'max_depth': 8, 
 'min_data_in_leaf': 202, 
 'learning_rate': 0.02361306858113298, 
 'feature_fraction': 0.7211579800007945, 
 'feature_fraction_bynode': 0.9081132434036039, 
 'bagging_fraction': 0.8727967232517579, 
 'bagging_freq': 7, 
 'lambda_l1': 5.763583420124947, 
 'lambda_l2': 14.460054311753241, 
 'min_gain_to_split': 0.032912218537174215, 
 'max_bin': 127}
'''
