==> notebooks/05_retrain_data_prep.ipynb <==
# Retrain - Data Preparation

1. Fit feature encoders to the whole training (train+valid) dataset and transform training (train+valid split)/test dataset
2. Retrain final model on the transformed training dataset

import pandas as pd
import numpy as np
from sklearn.preprocessing import OrdinalEncoder, OneHotEncoder

import warnings
warnings.simplefilter("ignore", category=pd.errors.PerformanceWarning)   # For convinience

# Load data
train_df = pd.read_csv('../data/processed/cleaned_train.csv')
valid_df = pd.read_csv('../data/processed/cleaned_valid.csv')
test_df = pd.read_csv('../data/processed/test_split.csv')
k_test_df = pd.read_csv('../data/raw/test.csv') # kaggle test set

# Combine train and valid into full training set
full_train_df = pd.concat([train_df, valid_df], ignore_index=True)
full_train_df

cat_vars = [var for var in full_train_df.columns if var.startswith('cat')]
cont_vars = [var for var in full_train_df.columns if var.startswith('cont')]

# Summarize all categorical columns
full_summary_cat_df = pd.DataFrame({
    "n_cats": full_train_df[cat_vars].nunique(),
    "most_freq_cat": full_train_df[cat_vars].agg(lambda x: x.value_counts().idxmax()),
    "most_freq_count": full_train_df[cat_vars].agg(lambda x: x.value_counts().max())
})

#full_summary_cat_df

# Fequency encoding
for cat in cat_vars:
    
    count = full_train_df[cat].value_counts()
    
    full_train_df[f'{cat}_freq'] = full_train_df[cat].map(count)
    full_train_df[f'{cat}_log_freq'] = np.log1p(full_train_df[f'{cat}_freq'])     # take log to compress extreme counts
    full_train_df[f'{cat}_norm_freq'] = full_train_df[f'{cat}_freq'] / len(full_train_df)

    test_df[f'{cat}_freq'] = test_df[cat].map(count).fillna(0)
    test_df[f'{cat}_log_freq'] = np.log1p(test_df[f'{cat}_freq'])     # take log to compress extreme counts
    test_df[f'{cat}_norm_freq'] = test_df[f'{cat}_freq'] / len(full_train_df)

    k_test_df[f'{cat}_freq'] = k_test_df[cat].map(count).fillna(0)
    k_test_df[f'{cat}_log_freq'] = np.log1p(k_test_df[f'{cat}_freq'])     # take log to compress extreme counts
    k_test_df[f'{cat}_norm_freq'] = k_test_df[f'{cat}_freq'] / len(full_train_df)

print('Train shape : ', full_train_df.shape)
print('Test shape  : ', test_df.shape)
print('Kaggle Test shape  : ', k_test_df.shape)

# Ordinal Encoding (Much easier to handle unseen data)
ord_en = OrdinalEncoder(handle_unknown='use_encoded_value', unknown_value=-1)
ord_en.fit(full_train_df[cat_vars])
    
full_train_df[cat_vars] = ord_en.transform(full_train_df[cat_vars])
test_df[cat_vars] = ord_en.transform(test_df[cat_vars])
k_test_df[cat_vars] = ord_en.transform(k_test_df[cat_vars])

print('Train shape : ', full_train_df.shape)
print('Test shape  : ', test_df.shape)
print('Kaggle Test shape  : ', k_test_df.shape)

# One Hot Encoding
# Only on low cardinality columns
low_card_cols = full_summary_cat_df[full_summary_cat_df['n_cats'] <= 3].index.to_list()

onehot_en = OneHotEncoder(handle_unknown='ignore', sparse_output=False)
onehot_en.fit(full_train_df[low_card_cols])

train_onehot_encoded = onehot_en.transform(full_train_df[low_card_cols])
test_onehot_encoded = onehot_en.transform(test_df[low_card_cols])
k_test_onehot_encoded = onehot_en.transform(k_test_df[low_card_cols])

train_onehot_df = pd.DataFrame(train_onehot_encoded, columns=onehot_en.get_feature_names_out(low_card_cols))
test_onehot_df = pd.DataFrame(test_onehot_encoded, columns=onehot_en.get_feature_names_out(low_card_cols))
k_test_onehot_df = pd.DataFrame(k_test_onehot_encoded, columns=onehot_en.get_feature_names_out(low_card_cols))

full_train_df = pd.concat([full_train_df, train_onehot_df], axis=1)
test_df = pd.concat([test_df, test_onehot_df], axis=1)
k_test_df = pd.concat([k_test_df, k_test_onehot_df], axis=1)

print('Train shape : ', full_train_df.shape)
print('Test shape  : ', test_df.shape)
print('Kaggle Test shape  : ', k_test_df.shape)

# Categorical groups numeric stats
for cat in cat_vars:
    group = full_train_df.groupby(cat)
    for cont in cont_vars:
        # Calculate global stats for unseen categories
        global_mean = full_train_df[cont].mean()
        global_med = full_train_df[cont].median()
        global_std = full_train_df[cont].std()
        
        full_train_df[f'{cat}_{cont}_mean'] = full_train_df[cat].map(group[cont].mean())
        test_df[f'{cat}_{cont}_mean'] = test_df[cat].map(group[cont].mean()).fillna(global_mean)
        k_test_df[f'{cat}_{cont}_mean'] = k_test_df[cat].map(group[cont].mean()).fillna(global_mean)

        full_train_df[f'{cat}_{cont}_med'] = full_train_df[cat].map(group[cont].median())
        test_df[f'{cat}_{cont}_med'] = test_df[cat].map(group[cont].median()).fillna(global_med)
        k_test_df[f'{cat}_{cont}_med'] = k_test_df[cat].map(group[cont].median()).fillna(global_med)

        full_train_df[f'{cat}_{cont}_std'] = full_train_df[cat].map(group[cont].std())
        test_df[f'{cat}_{cont}_std'] = test_df[cat].map(group[cont].std()).fillna(global_std)
        k_test_df[f'{cat}_{cont}_std'] = k_test_df[cat].map(group[cont].std()).fillna(global_std)

print('Train shape : ', full_train_df.shape)
print('Test shape  : ', test_df.shape)
print('Kaggle Test shape  : ', k_test_df.shape)

# Numerics transformation

# Rank encoding for numerics
def fit_rank_transform(df, col):
    unique = sorted(df[col].unique())
    ranks = np.searchsorted(unique, unique) / len(unique)
    return unique, ranks

def transform_rank(values, unique, ranks):
    ''' Return the rank of value given the fitted rank transform (unique, ranks) '''
    return np.interp(values, unique, ranks)

# Winsorization (cap extremes)
def fit_winsor(df, col):
    l, u = df[col].quantile([0.01, 0.99])
    return l, u


for cont in cont_vars:
    # Log transform
    full_train_df[f'log_{cont}'] = np.log1p(full_train_df[cont])
    test_df[f'log_{cont}'] = np.log1p(test_df[cont])
    k_test_df[f'log_{cont}'] = np.log1p(k_test_df[cont])

    # Rank transform
    uniq, ranks = fit_rank_transform(full_train_df, cont)
    full_train_df[f'{cont}_rank'] = transform_rank(full_train_df[cont], uniq, ranks)
    test_df[f'{cont}_rank'] = transform_rank(test_df[cont], uniq, ranks)
    k_test_df[f'{cont}_rank'] = transform_rank(k_test_df[cont], uniq, ranks)

    # Winsorization
    l, u = fit_winsor(full_train_df, cont)
    full_train_df[f'{cont}_cap'] = full_train_df[cont].clip(l, u)
    test_df[f'{cont}_cap'] = test_df[cont].clip(l, u)
    k_test_df[f'{cont}_cap'] = k_test_df[cont].clip(l, u)

# Transform target loss
full_train_df[f'log_loss'] = np.log(full_train_df['loss'])
test_df[f'log_loss'] = np.log(test_df['loss'])

print('Train shape : ', full_train_df.shape)
print('Test shape  : ', test_df.shape)
print('Kaggle Test shape  : ', k_test_df.shape)

# Export transformed retrain data
full_train_df.to_parquet('../data/retrain/final_transformed_train.parquet', index=False)
test_df.to_parquet('../data/retrain/final_transformed_test.parquet', index=False)
k_test_df.to_parquet('../data/retrain/final_k_test.parquet', index=False)
