==> notebooks/06_final_retrain.ipynb <==
# Retrain Final Model & Evaluation

import pandas as pd
import numpy as np
import lightgbm as lgb
from sklearn.metrics import mean_absolute_error

# Load retrain data
train_df = pd.read_parquet('../data/retrain/final_train.parquet')
test_df = pd.read_parquet('../data/retrain/final_test.parquet')

# Prepare data for lightgbm
X_train, Y_train = train_df.drop(columns=['loss', 'log_loss']), train_df['log_loss']

train_lgb = lgb.Dataset(X_train, label=Y_train)

params = {
    'objective': 'regression',
    'metric': 'mae',
    'num_leaves': 404,
    'max_depth': 8, 
    'min_data_in_leaf': 202, 
    'learning_rate': 0.02361306858113298, 
    'feature_fraction': 0.7211579800007945, 
    'feature_fraction_bynode': 0.9081132434036039, 
    'bagging_fraction': 0.8727967232517579, 
    'bagging_freq': 7, 
    'lambda_l1': 5.763583420124947, 
    'lambda_l2': 14.460054311753241, 
    'min_gain_to_split': 0.03291221t8537174215, 
    'max_bin': 127,
    'boosting_type': 'gbdt',
    'seed': 42,
}

# Train model
model = lgb.train(
    params,
    train_lgb,
    num_boost_round=1400   # Use best_iterations for this set of params during tuning phase
)

test_df

## Evaluation on test split

loss_true = test_df['loss']
loss_pred = np.exp(model.predict(test_df.drop(columns=['loss','log_loss'])))
mae = mean_absolute_error(loss_true, loss_pred)
print('MAE : ', mae)

# Export model
model.save_model('final_model.txt')
